+++ CHANGELOG.md
@@ -23,6 +23,7 @@
 * activating extensions can now take an options hash
 * Don't re-minify files with ".min" in their name
 * Serve purely static folders directly (without source/ and config.rb)
+* Set ignored files and disable directory_indexes from YAML frontmatter
 
 2.0.14
 ====